Moved RPCContainer to a separate file. More serialization code
